Refactor shared SingleReduction code into a base algorithm class
